requirements.txt
Django==3.0.8
django-crispy-forms==1.9.1
djangorestframework==3.11.0
Pillow==7.1.2
opencv-python
